refactor: move and rename game.php
